update - move keyboard in its folder for further improvement to come
